+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Deprecates CrashReporting.reportJSException in favour of a new strongly typed API: CrashReporting.reportError
 - Fixes Survey interface export causing a build error with certain babel versions
 
 ## 11.5.0 (2022-11-28)